version2/backend/requirements.txt
annotated-types==0.7.0
email-validator==2.1.1
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
joblib==1.5.2
motor==3.7.1
narwhals==2.5.0
numpy==2.3.3
openai==1.109.1
openpyxl==3.1.2
packaging==25.0
pandas==2.3.2
passlib==1.7.4
plotly==6.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
xlrd==2.0.1
aiofiles==24.1.0
python-magic==0.4.27
